+++ server/tests/backends/mongo_translator/test_mongo_translator_steps.py
@@ -1,13 +1,16 @@
 import json
 import socket
 import uuid
+from collections.abc import Generator
 from io import StringIO
 
 import docker
+import geopandas as gpd
 import pandas as pd
 import pytest
 from pandas.api.types import is_bool_dtype, is_datetime64_any_dtype, is_numeric_dtype
 from pymongo import MongoClient
+from pymongo.database import Database
 
 from tests.utils import assert_dataframes_content_equals, get_spec_from_json_fixture, retrieve_case
 from weaverbird.backends.mongo_translator.mongo_pipeline_translator import translate_pipeline
@@ -18,7 +21,7 @@ exec_type = "mongo"
 test_cases = retrieve_case("mongo_translator", exec_type)
 
 
-def unused_port():
+def unused_port() -> int:
     """find an unused TCP port and return it"""
     with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
         s.bind(("127.0.0.1", 0))
@@ -26,12 +29,12 @@ def unused_port():
 
 
 @pytest.fixture(scope="session")
-def mongo_version():
+def mongo_version() -> str:
     return "7"
 
 
 @pytest.fixture(scope="session")
-def mongo_server_port(mongo_version):
+def mongo_server_port(mongo_version: int) -> Generator[int, None, None]:
     port = unused_port()
     docker_client = docker.from_env()
     container = docker_client.containers.run(
@@ -42,9 +45,19 @@ def mongo_server_port(mongo_version):
 
 
 @pytest.fixture()
-def mongo_database(mongo_server_port):
+def mongo_database(mongo_server_port: int, case_id: str) -> Database:
     client = MongoClient("localhost", mongo_server_port)
-    return client["tests"]
+    # Database names cannot contain the character '/'
+    # db name must be at most 63 characters
+    return client[case_id[1:63]]
+
+
+def _serialize_geo_df(gdf: pd.DataFrame) -> pd.DataFrame:
+    """Converting the GeoSeries in shapely format to GeoJSON features"""
+    gdf = gpd.GeoDataFrame(gdf)
+    geo_series = gdf.geometry
+    gdf[geo_series.name] = json.loads(geo_series.to_json())["features"]
+    return pd.DataFrame(gdf)
 
 
 def _sanitize_df(df: pd.DataFrame) -> pd.DataFrame:
@@ -69,8 +82,18 @@ def _sanitize_df(df: pd.DataFrame) -> pd.DataFrame:
     return df[sorted(df.columns)]  # order of columns may be different between pandas and mongo
 
 
+def _load_df(spec: dict) -> pd.DataFrame:
+    return (
+        pd.DataFrame(gpd.read_file(json.dumps(spec["data"])))
+        if spec.get("schema") == "geojson"
+        else pd.read_json(StringIO(json.dumps(spec)), orient="table")
+    )
+
+
 def _sanitized_df_from_pandas_table(df_spec: dict) -> pd.DataFrame:
-    df = _sanitize_df(pd.read_json(StringIO(json.dumps(df_spec)), orient="table"))
+    df = _sanitize_df(_load_df(df_spec))
+    if df_spec["schema"] == "geojson":
+        return df
     bool_cols = [f["name"] for f in df_spec["schema"]["fields"] if f["type"] == "boolean"]
     # By default, pandas converts null bools to False. This enforces
     # use of nullable booleans
@@ -80,19 +103,25 @@ def _sanitized_df_from_pandas_table(df_spec: dict) -> pd.DataFrame:
     return df
 
 
+def insert_data_to_mongo(mongo_database, collection_uid: str, schema: dict | str, df: pd.DataFrame):
+    if schema == "geojson":
+        df = _serialize_geo_df(df)
+    mongo_database[collection_uid].insert_many(df.to_dict(orient="records"))
+
+
 @pytest.mark.parametrize("case_id,case_spec_file_path", test_cases)
 def test_mongo_translator_pipeline(mongo_database, case_id, case_spec_file_path, available_variables):
     # insert in mongoDB
     collection_uid = uuid.uuid4().hex
     spec = get_spec_from_json_fixture(case_id, case_spec_file_path)
-    data = _sanitized_df_from_pandas_table(spec["input"]).to_dict(orient="records")
-    mongo_database[collection_uid].insert_many(data)
+    data = _sanitized_df_from_pandas_table(spec["input"])
+    insert_data_to_mongo(mongo_database, collection_uid, spec["input"]["schema"], data)
     if "other_inputs" in spec and (
         "join" in case_id or "append" in case_id
     ):  # needed for join & append steps tests as we need a != collection
         for collection_name, raw_df in spec["other_inputs"].items():
-            df = pd.read_json(StringIO(json.dumps(raw_df)), orient="table")
-            mongo_database[collection_name].insert_many(df.to_dict(orient="records"))
+            df = _load_df(raw_df)
+            insert_data_to_mongo(mongo_database, collection_name, raw_df["schema"], df)
 
     # create query
     steps = spec["step"]["pipeline"]